+++ packages/web/src/app/clubs/[id]/page.tsx
@@ -4,9 +4,9 @@ import { useParams } from "next/navigation";
 
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
-import ClubDetailPublicFrame from "@sparcs-clubs/web/features/clubDetails/frames/ClubDetailPublicFrame";
-import ClubDetailStudentFrame from "@sparcs-clubs/web/features/clubDetails/frames/ClubDetailStudentFrame";
-import { useGetClubDetail } from "@sparcs-clubs/web/features/clubDetails/services/getClubDetail";
+import ClubDetailPublicFrame from "@sparcs-clubs/web/features/clubs/frames/ClubDetailPublicFrame";
+import ClubDetailStudentFrame from "@sparcs-clubs/web/features/clubs/frames/ClubDetailStudentFrame";
+import { useGetClubDetail } from "@sparcs-clubs/web/features/clubs/services/getClubDetail";
 import isStudent from "@sparcs-clubs/web/utils/isStudent";
 
 const ClubDetail = () => {
