+++ packages/web/src/features/clubs/components/ClubDetailCard.tsx
