+++ packages/web/src/features/clubs/services/_mock/mockupClubDetail.ts
