+++ packages/web/src/features/my/frames/_atomic/MyMemberRegisterFrame.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
 import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
 import MoreDetailTitle from "@sparcs-clubs/web/common/components/MoreDetailTitle";
-import { useGetMyMemberRegistration } from "@sparcs-clubs/web/features/clubDetails/services/getMyMemberRegistration";
+import { useGetMyMemberRegistration } from "@sparcs-clubs/web/features/clubs/services/getMyMemberRegistration";
 import MyMemberTable from "@sparcs-clubs/web/features/my/components/MyMemberTable";
 
 const MyMemberRegisterFrame: React.FC = () => {
